+++ mw/preference.ts
@@ -1,4 +1,4 @@
-import {rules} from 'wikiparser-node/dist/base';
+import {rules} from 'wikiparser-node/base';
 import {CodeMirror} from './base';
 import {getObject, setObject} from './util';
 import {msg, parseMsg, i18n} from './msg';
@@ -38,7 +38,7 @@ const enum RuleState {
 }
 
 export const indentKey = 'codemirror-mediawiki-indent',
-	prefs = new Set<string>(getObject(storageKey) as string[] | null),
+	prefs = new Set(getObject(storageKey) as string[] | null),
 	wikilintConfig = (getObject(wikilintKey) || {}) as Record<LintError.Rule, RuleState | undefined>,
 	codeConfigs = new Map(codeKeys.map(k => [k, getObject(`codemirror-mediawiki-${k}`)]));
 
@@ -79,7 +79,7 @@ export const loadJSON = (async () => {
 	const params: ApiQueryRevisionsParams = {
 		action: 'query',
 		prop: 'revisions',
-		titles: userPage,
+		titles: userPage!,
 		rvprop: 'content',
 		rvlimit: 1,
 	};
@@ -260,7 +260,7 @@ export const openPreference = async (editors: (CodeMirror | undefined)[]): Promi
 		if (changed && user && (save || prefs.has('save'))) {
 			const params: ApiEditPageParams = {
 				action: 'edit',
-				title: userPage,
+				title: userPage!,
 				text: JSON.stringify({
 					addons: [...prefs],
 					indent,