+++ src/global.d.ts
@@ -1,13 +1,8 @@
 import {CodeMirror6} from './codemirror';
-import type {Config, Rule} from 'wikiparser-node';
+import 'wikiparser-node/extensions/typings';
 import type {Linter} from 'eslint';
 import type {PublicApi} from 'stylelint';
-import type {Diagnostic} from '@codemirror/lint';
 import type {MwConfig, LintSource} from './codemirror';
-
-class WikiLinter {
-	codemirror(s: string): Promise<(Diagnostic & {rule: Rule})[]>;
-}
 
 interface luaparse {
 	defaultOptions: {luaVersion: string};
@@ -25,12 +20,6 @@ declare global {
 		export {CodeMirror6, CDN};
 	}
 
-	const wikiparse: {
-		getConfig(): Promise<Config>;
-		setConfig(config: Config): void;
-		setI18N(i18n: Record<string, string>): void;
-		Linter: new (include?: boolean) => WikiLinter;
-	};
 	const eslint: {
 		Linter: new () => Linter;
 	};